Moved Mock URL Handler to it's own file and Added test for ForgettenPass
